refactor(init): split init.rs into the five things it does
A 2,500-line module held five separable concerns with no boundary between
them: where an installation puts its files, what the policy file's lifecycle
is, everything that goes through the `claude` CLI, who owns the runtime
endpoint, and what the run reports at the end.

`init/mod.rs` keeps the sequence itself — `claude_code`, `switch_over`,
`Compensation`/`Undo`, the launcher and the binary install — and the four
concerns move to `paths`, `config`, `claude`, `endpoint` and `receipt`.

Function bodies move unchanged. The only edits are `pub(super)` on the
thirty-five items a sibling names, the `use` lines each module needs, and
`pub use self::paths::installed_config_path` so the crate's surface does not
move. Verified mechanically: all 92 function bodies compare byte-identical to
their originals once visibility qualifiers and indentation are normalized.

Tests move with the code they exercise, so each module's cases sit beside it
rather than in one 500-line block.
